# Optimize Vector.Max codegen using AVX512