--- a/src/coreclr/jit/gentree.cpp
+++ b/src/coreclr/jit/gentree.cpp
@@ -24246,29 +24246,28 @@ GenTree* Compiler::gtNewSimdMaxNode(
             // We still need to check for NaN as vrangeps/vrangepd does not handle NaN as specified in IEEE 754 2019.
             //
             // This can be represented as the following managed code:
-            //      Vector128<float> nanMask = Vector128.IsNaN(op1) | Vector128.IsNaN(op2);
-            //      Vector128<float> rangeResult = Avx512DQ.VL.Range(op1, op2, 0x5);
-            //      return Vector128.ConditionalSelect(nanMask, Vector128.Create(float.NaN), rangeResult);
+            //      Vector128<float> range = Avx512DQ.VL.Range(op1, op2, 0x5);
+            //      Vector128<float> fixup1 = Avx512F.VL.Fixup(op1, op2, Vector128<int>.One, 0);
+            //      Vector128<float> fixup2 = Avx512F.VL.Fixup(range, fixup1, Vector128<int>.One, 0);
+            //      return fixup2
             //
-            // 0x5 is the control byte:
+            // 0x5 is the control byte for vrangeps/vrangepd:
             //      Imm8[1:0] = 01b : Select Max value
             //      Imm8[3:2] = 01b : Select sign(Compare_Result)
 
-            GenTree* op1Dup1    = fgMakeMultiUse(&op1);
-            GenTree* op2Dup1    = fgMakeMultiUse(&op2);
-            GenTree* op1NaNMask = gtNewSimdIsNaNNode(type, op1, simdBaseJitType, simdSize);
-            GenTree* op2NaNMask = gtNewSimdIsNaNNode(type, op2, simdBaseJitType, simdSize);
-            GenTree* nanMask    = gtNewSimdBinOpNode(GT_OR, type, op1NaNMask, op2NaNMask, simdBaseJitType, simdSize);
-            GenTree* rangeOp    = gtNewSimdHWIntrinsicNode(type, op1Dup1, op2Dup1, gtNewIconNode(0x5),
+            GenTree* op1Dup     = fgMakeMultiUse(&op1);
+            GenTree* op2Dup     = fgMakeMultiUse(&op2);
+            GenTree* rangeOp    = gtNewSimdHWIntrinsicNode(type, op1, op2, gtNewIconNode(0x5),
                                                         simdSize == 64 ? NI_AVX512DQ_Range : NI_AVX512DQ_VL_Range,
                                                            simdBaseJitType, simdSize);
-            GenTreeVecCon* vecCon = gtNewVconNode(type);
+            GenTreeVecCon* tblVecCon1 = gtNewVconNode(type);
+            GenTreeVecCon* tblVecCon2 = gtNewVconNode(type);
 
             if (simdBaseType == TYP_FLOAT)
             {
                 for (unsigned i = 0; i < (simdSize / sizeof(float)); i++)
                 {
-                    vecCon->gtSimdVal.f32[i] = nanf("");
+                    tblVecCon1->gtSimdVal.i32[i] = tblVecCon2->gtSimdVal.i32[i] = 0x1;
                 }
             }
             else
@@ -24276,10 +24275,16 @@ GenTree* Compiler::gtNewSimdMaxNode(
                 assert(simdBaseType == TYP_DOUBLE);
                 for (unsigned i = 0; i < (simdSize / sizeof(double)); i++)
                 {
-                    vecCon->gtSimdVal.f64[i] = nan("");
+                    tblVecCon1->gtSimdVal.i64[i] = tblVecCon2->gtSimdVal.i64[i] = 0x1;
                 }
             }
-            return gtNewSimdCndSelNode(type, nanMask, vecCon, rangeOp, simdBaseJitType, simdSize);
+
+            NamedIntrinsic fixupIntrinsic = simdSize == 64 ? NI_AVX512F_Fixup : NI_AVX512F_VL_Fixup;
+            GenTree*       fixup1         = gtNewSimdHWIntrinsicNode(type, op1Dup, op2Dup, tblVecCon1, gtNewIconNode(0),
+                                                                     fixupIntrinsic, simdBaseJitType, simdSize);
+            GenTree*       fixup2         = gtNewSimdHWIntrinsicNode(type, rangeOp, fixup1, tblVecCon2, gtNewIconNode(0),
+                                                                     fixupIntrinsic, simdBaseJitType, simdSize);
+            return fixup2;
         }