This PR optimizes the codegen for `Vector128/256/512.Max` when AVX512 is supported. The new codegen emitted is the same pattern as the scalar codegen for `Math.Max`.

<details>

<summary>Scalar version + codegen</summary>

```csharp
public double ScalarMax(double x, double y) {
        return Math.Max(x,y);
}
```

```
C.ScalarMax(Double, Double)
    L0000: vrangesd xmm0, xmm1, xmm2, 5
    L0007: vmovups xmm3, [0x7ffcb40e0060]
    L000f: vfixupimmsd xmm1, xmm2, xmm3, 0
    L0016: vfixupimmsd xmm0, xmm1, xmm3, 0
    L001d: ret
```

</details>

<details>

<summary>Vector version + codegen</summary>

```
public Vector128<double> VectorMax(Vector128<double> x, Vector128<double> y) {
        return Vector128.Max(x, y);
}
```

Codegen (before):
```
       vmovups   xmm0,[r8]
       vmovups   xmm1,[r9]
       vcmpeqpd  xmm2,xmm1,xmm0
       vxorps    xmm3,xmm3,xmm3
       vpcmpgtq  xmm3,xmm3,xmm1
       vcmpneqpd xmm4,xmm0,xmm0
       vpternlogq xmm4,xmm3,xmm2,0F8
       vcmpltpd  xmm2,xmm1,xmm0
       vorpd     xmm2,xmm2,xmm4
       vpternlogq xmm2,xmm1,xmm0,0AC
       vmovups   [rdx],xmm2
       mov       rax,rdx
       ret
```

Codegen (after):
```
       vmovups   xmm0,[r8]
       vmovups   xmm1,[r9]
       vrangepd  xmm2,xmm0,xmm1,5
       vmovddup  xmm3,qword ptr [7FFA5DFC3030]
       vfixupimmpd xmm0,xmm1,xmm3,0
       vfixupimmpd xmm2,xmm0,xmm3,0
       vmovups   [rdx],xmm2
       mov       rax,rdx
       ret
```

</details>

<details>

<summary>Performance</summary>

```
| Type                                | Method     | Job        | Toolchain                              | BufferLength | Mean       | Error      | StdDev     | Median     | Min        | Max        | Ratio | RatioSD | Allocated | Alloc Ratio |
|------------------------------------ |----------- |----------- |--------------------------------------- |------------- |-----------:|-----------:|-----------:|-----------:|-----------:|-----------:|------:|--------:|----------:|------------:|
| Perf_NumberTensorPrimitives<Double> | Max_Vector | Job-DWINXP | \core_roots\base\Core_Root\corerun.exe | 128          |  23.580 ns |  0.4058 ns |  0.3597 ns |  23.672 ns |  22.363 ns |  23.797 ns |  1.00 |    0.02 |         - |          NA |
| Perf_NumberTensorPrimitives<Double> | Max_Vector | Job-OMRYDZ | \core_roots\diff\Core_Root\corerun.exe | 128          |  15.917 ns |  0.0598 ns |  0.0530 ns |  15.912 ns |  15.797 ns |  16.002 ns |  0.68 |    0.01 |         - |          NA |
|                                     |            |            |                                        |              |            |            |            |            |            |            |       |         |           |             |
| Perf_NumberTensorPrimitives<Single> | Max_Vector | Job-DWINXP | \core_roots\base\Core_Root\corerun.exe | 128          |  13.422 ns |  0.1850 ns |  0.1731 ns |  13.425 ns |  13.119 ns |  13.691 ns |  1.00 |    0.02 |         - |          NA |
| Perf_NumberTensorPrimitives<Single> | Max_Vector | Job-OMRYDZ | \core_roots\diff\Core_Root\corerun.exe | 128          |   9.448 ns |  0.1728 ns |  0.1617 ns |   9.427 ns |   9.265 ns |   9.781 ns |  0.70 |    0.01 |         - |          NA |
|                                     |            |            |                                        |              |            |            |            |            |            |            |       |         |           |             |
| Perf_NumberTensorPrimitives<Double> | Max_Scalar | Job-DWINXP | \core_roots\base\Core_Root\corerun.exe | 128          |  17.686 ns |  0.2818 ns |  0.2498 ns |  17.797 ns |  17.083 ns |  17.856 ns |  1.00 |    0.02 |         - |          NA |
| Perf_NumberTensorPrimitives<Double> | Max_Scalar | Job-OMRYDZ | \core_roots\diff\Core_Root\corerun.exe | 128          |  10.053 ns |  0.0393 ns |  0.0368 ns |  10.060 ns |   9.986 ns |  10.117 ns |  0.57 |    0.01 |         - |          NA |
|                                     |            |            |                                        |              |            |            |            |            |            |            |       |         |           |             |
| Perf_NumberTensorPrimitives<Single> | Max_Scalar | Job-DWINXP | \core_roots\base\Core_Root\corerun.exe | 128          |  10.899 ns |  0.1788 ns |  0.1493 ns |  10.944 ns |  10.412 ns |  10.995 ns |  1.00 |    0.02 |         - |          NA |
| Perf_NumberTensorPrimitives<Single> | Max_Scalar | Job-OMRYDZ | \core_roots\diff\Core_Root\corerun.exe | 128          |   6.588 ns |  0.0550 ns |  0.0515 ns |   6.576 ns |   6.492 ns |   6.664 ns |  0.60 |    0.01 |         - |          NA |
|                                     |            |            |                                        |              |            |            |            |            |            |            |       |         |           |             |
| Perf_NumberTensorPrimitives<Double> | Max        | Job-DWINXP | \core_roots\base\Core_Root\corerun.exe | 128          |  39.597 ns |  0.3459 ns |  0.3236 ns |  39.617 ns |  38.713 ns |  40.068 ns |  1.00 |    0.01 |         - |          NA |
| Perf_NumberTensorPrimitives<Double> | Max        | Job-OMRYDZ | \core_roots\diff\Core_Root\corerun.exe | 128          |  20.630 ns |  0.0695 ns |  0.0616 ns |  20.622 ns |  20.556 ns |  20.759 ns |  0.52 |    0.00 |         - |          NA |
|                                     |            |            |                                        |              |            |            |            |            |            |            |       |         |           |             |
| Perf_NumberTensorPrimitives<Single> | Max        | Job-DWINXP | \core_roots\base\Core_Root\corerun.exe | 128          |  23.512 ns |  0.0944 ns |  0.0883 ns |  23.520 ns |  23.288 ns |  23.653 ns |  1.00 |    0.01 |         - |          NA |
| Perf_NumberTensorPrimitives<Single> | Max        | Job-OMRYDZ | \core_roots\diff\Core_Root\corerun.exe | 128          |  11.249 ns |  0.1301 ns |  0.1217 ns |  11.295 ns |  10.971 ns |  11.385 ns |  0.48 |    0.01 |         - |          NA |
|                                     |            |            |                                        |              |            |            |            |            |            |            |       |         |           |             |
| Perf_NumberTensorPrimitives<Double> | Max_Vector | Job-DWINXP | \core_roots\base\Core_Root\corerun.exe | 3079         | 549.776 ns |  3.1703 ns |  2.9655 ns | 550.484 ns | 544.151 ns | 553.314 ns |  1.00 |    0.01 |         - |          NA |
| Perf_NumberTensorPrimitives<Double> | Max_Vector | Job-OMRYDZ | \core_roots\diff\Core_Root\corerun.exe | 3079         | 441.097 ns |  1.1802 ns |  1.1039 ns | 441.030 ns | 439.368 ns | 442.973 ns |  0.80 |    0.00 |         - |          NA |
|                                     |            |            |                                        |              |            |            |            |            |            |            |       |         |           |             |
| Perf_NumberTensorPrimitives<Single> | Max_Vector | Job-DWINXP | \core_roots\base\Core_Root\corerun.exe | 3079         | 335.473 ns |  4.3674 ns |  4.0853 ns | 336.539 ns | 324.981 ns | 339.704 ns |  1.00 |    0.02 |         - |          NA |
| Perf_NumberTensorPrimitives<Single> | Max_Vector | Job-OMRYDZ | \core_roots\diff\Core_Root\corerun.exe | 3079         | 183.924 ns |  0.3314 ns |  0.2938 ns | 183.908 ns | 183.439 ns | 184.449 ns |  0.55 |    0.01 |         - |          NA |
|                                     |            |            |                                        |              |            |            |            |            |            |            |       |         |           |             |
| Perf_NumberTensorPrimitives<Double> | Max_Scalar | Job-DWINXP | \core_roots\base\Core_Root\corerun.exe | 3079         | 376.094 ns |  1.9991 ns |  1.8700 ns | 376.419 ns | 370.356 ns | 378.039 ns |  1.00 |    0.01 |         - |          NA |
| Perf_NumberTensorPrimitives<Double> | Max_Scalar | Job-OMRYDZ | \core_roots\diff\Core_Root\corerun.exe | 3079         | 303.696 ns |  1.1710 ns |  1.0954 ns | 303.673 ns | 302.163 ns | 305.591 ns |  0.81 |    0.00 |         - |          NA |
|                                     |            |            |                                        |              |            |            |            |            |            |            |       |         |           |             |
| Perf_NumberTensorPrimitives<Single> | Max_Scalar | Job-DWINXP | \core_roots\base\Core_Root\corerun.exe | 3079         | 177.732 ns |  0.6656 ns |  0.5558 ns | 177.806 ns | 176.879 ns | 178.902 ns |  1.00 |    0.00 |         - |          NA |
| Perf_NumberTensorPrimitives<Single> | Max_Scalar | Job-OMRYDZ | \core_roots\diff\Core_Root\corerun.exe | 3079         | 107.079 ns |  0.3884 ns |  0.3443 ns | 107.008 ns | 106.533 ns | 107.741 ns |  0.60 |    0.00 |         - |          NA |
|                                     |            |            |                                        |              |            |            |            |            |            |            |       |         |           |             |
| Perf_NumberTensorPrimitives<Double> | Max        | Job-DWINXP | \core_roots\base\Core_Root\corerun.exe | 3079         | 902.804 ns | 16.9216 ns | 16.6193 ns | 907.864 ns | 848.689 ns | 921.769 ns |  1.00 |    0.03 |         - |          NA |
| Perf_NumberTensorPrimitives<Double> | Max        | Job-OMRYDZ | \core_roots\diff\Core_Root\corerun.exe | 3079         | 513.187 ns |  2.4330 ns |  2.2758 ns | 513.181 ns | 510.050 ns | 517.755 ns |  0.57 |    0.01 |         - |          NA |
|                                     |            |            |                                        |              |            |            |            |            |            |            |       |         |           |             |
| Perf_NumberTensorPrimitives<Single> | Max        | Job-DWINXP | \core_roots\base\Core_Root\corerun.exe | 3079         | 461.626 ns |  3.2858 ns |  3.0735 ns | 462.374 ns | 453.206 ns | 465.063 ns |  1.00 |    0.01 |         - |          NA |
| Perf_NumberTensorPrimitives<Single> | Max        | Job-OMRYDZ | \core_roots\diff\Core_Root\corerun.exe | 3079         | 253.227 ns |  1.0693 ns |  1.0003 ns | 253.197 ns | 251.372 ns | 254.555 ns |  0.55 |    0.00 |         - |          NA |
```

</details>