@EgorBo 2025-06-06
These Min and Max routine look like the can be unified to remove the code duplication: https://www.diffchecker.com/SYTwWTzO/