@tannergooding 2025-06-06
> These Min and Max routines look like they can be unified to remove the code duplication:

Right. They are unified for the scalar path already too.

I think when we finish the factoring to also accelerate `*Magnitude`, `*MagnitudeNumber`, and `*Number`, we should do so by extracting the scalar path into a reusable helper that takes 1 extra `bool isScalar` parameter. That way we can have one function that handles generation for all 16 variations (8 scalar, 8 vector).