@alexcovington 2025-06-06
> @alexcovington is also covering the other Min/Max APIs listed something you'd be interested in doing? If not, I can get something up over the weekend handling them.

I'm not sure it's something I can commit to right now. It is probably better to assign the other variants to someone else for now.